% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactCheckbox.R
\name{checkbox}
\alias{checkbox}
\title{Define a checkbox}
\usage{
checkbox(label, value, class = "")
}
\arguments{
\item{label}{the label of the checkbox, can be an ordinary character string,
a \code{html} character string made with the \code{\link[htmltools]{HTML}}
function, or a \code{shiny.tag} object}

\item{value}{the initial value of the checkbox, \code{TRUE} (checked), or
\code{FALSE} (unchecked)}

\item{class}{the name of a CSS class to attach to the checkbox, that should
be defined with \code{\link{checkboxStyle}}}
}
\value{
A named list, to be used in \code{\link{reactCheckboxesInput}}.
}
\description{
Define a checkbox by its label, its value, and optionally a
  CSS class.
}
